+++ packages/datadog_flags/lib/datadog_flags.dart
@@ -13,6 +13,8 @@ library;
 export 'src/datadog_flags_config.dart'
     show DatadogFlagsConfig, DatadogFlagsSite;
 export 'src/datadog_flags.dart' show DatadogFlags;
+export 'src/assignment_request_client.dart'
+    show withAssignmentRequestRetry, withAssignmentRequestTimeout;
 export 'src/flags_client.dart' show DatadogFlagsClient, FlagDetails;
 export 'src/flags_configuration.dart' show DatadogFlagsConfiguration;
 export 'src/flags_error.dart' show FlagEvaluationError;